+++ package/Runtime/DeviceSensors/Scripts/Samples.Sensors.codegen/PanoramaControls.cs
@@ -7,21 +7,39 @@ namespace Needle.Typescript.GeneratedComponents
 {
 	public partial class PanoramaControls : UnityEngine.MonoBehaviour
 	{
-		public bool @gyroscopeInput = true;
+		public Needle.Typescript.GeneratedComponents.GyroscopeMode @gyroscopeMode;
 		public bool @pointerInput = true;
-		public float @rotateSpeed = 1f;
+		public float @rotateSpeed = 0.16f;
 		public float @rotateSmoothing = 5f;
 		public bool @enableZoom = true;
 		public float @zoomMin = 40f;
 		public float @zoomMax = 90f;
 		public float @zoomSpeed = 0.1f;
 		public float @zoomSmoothing = 5f;
-		public bool @autoRotate = true;
+		public Needle.Typescript.GeneratedComponents.AutoRotateMode @autoRotate;
 		public float @autoRotateSpeed = 0.15f;
 		public float @autoRotateTimeout = 3f;
 		public void start(){}
 		public void onBeforeRender(){}
+		public void setGyroscope(bool @state){}
 	}
 }
 
 // NEEDLE_CODEGEN_END
+
+namespace Needle.Typescript.GeneratedComponents
+{
+	public enum GyroscopeMode
+	{
+		Disabled = 0,
+		ControlledViaButton = 1,
+		Enabled
+	}
+
+    public enum AutoRotateMode
+    {
+        Disabled = 0,
+        OnlyOnceOnStart = 1,
+        Enabled = 2
+    }
+}