+++ package/Runtime/Facefilter/FaceFilter.codegen/ShaderToyFaceFilter.cs
@@ -1,0 +1,17 @@
+// NEEDLE_CODEGEN_START
+// auto generated code - do not edit directly
+
+#pragma warning disable
+
+namespace Needle.Typescript.GeneratedComponents
+{
+	public partial class ShaderToyFaceFilter : Needle.Typescript.GeneratedComponents.FaceMeshBehaviour
+	{
+		public UnityEngine.Texture @mask;
+		public void update(){}
+		public void OnEnable(){}
+		public void OnDisable(){}
+	}
+}
+
+// NEEDLE_CODEGEN_END