+++ package/Runtime/_Shared/SampleScripts.codegen/ConsumeAllPointers.cs
@@ -1,0 +1,25 @@
+// NEEDLE_CODEGEN_START
+// auto generated code - do not edit directly
+
+#pragma warning disable
+
+namespace Needle.Typescript.GeneratedComponents
+{
+	public partial class ConsumeAllPointers : UnityEngine.MonoBehaviour
+	{
+		public bool @consumeClick = true;
+		public bool @consumeDown = true;
+		public bool @consumeMove = true;
+		public bool @consumeUp = true;
+		public bool @consumeHover = true;
+		public bool @stopPropagation = true;
+		public void onPointerDown(UnityEngine.EventSystems.PointerEventData @event){}
+		public void onPointerMove(UnityEngine.EventSystems.PointerEventData @event){}
+		public void onPointerUp(UnityEngine.EventSystems.PointerEventData @event){}
+		public void onPointerClick(UnityEngine.EventSystems.PointerEventData @event){}
+		public void onPointerEnter(UnityEngine.EventSystems.PointerEventData @event){}
+		public void onPointerExit(UnityEngine.EventSystems.PointerEventData @event){}
+	}
+}
+
+// NEEDLE_CODEGEN_END