+++ package/Runtime/_Shared/SampleScripts.codegen/LODGroupUI.cs
@@ -1,0 +1,16 @@
+// NEEDLE_CODEGEN_START
+// auto generated code - do not edit directly
+
+#pragma warning disable
+
+namespace Needle.Typescript.GeneratedComponents
+{
+	public partial class LODGroupUI : UnityEngine.MonoBehaviour
+	{
+		public UnityEngine.LODGroup @lodGroup;
+		public UnityEngine.UI.Text @text;
+		public void update(){}
+	}
+}
+
+// NEEDLE_CODEGEN_END